Create cats list and modify script for list
Stored the values within a list so that the code can use a loop to
execute same code over for both cats. This way the code is concise
and efficient. It would help if we were to add in more cats into the
list and we wouldn't have to add in any more code for it.