Add clicks to each cat
Add in a heading that would show the name and number of clicks for each
cat. Had to add heading to the top to show the name of the cat with the
number of clicks that cat has. Matched it with the layout of the images
so the user can see which cat is which.