Add buttons for each cat
Temporarily removed the display and click count of each cat. Add in a
list of cat on the top of the page with all the cats' names that will
later display the selected cats image and click count.

I put the list as a button to indicate that the user can click on the
names to select which cat to be displayed.