Organize code using MVO
This is so we can keep our code organized and scalable. I kept the
model, view, and octopus all separate. The model contains the database
information, the view takes care of the display, and the octopus is
is the connection between the model and the view. This will help for
future changes to the script so it will be easier to modify.