Change to use KnockoutJS
This eliminates the functionality that would grab the data from the
model and put it in the view because of the synchronous updates that
Knockout offers. I applied the bindings in the index and added in the
viewmodel in the app.js file which is more efficient than the last MVO
structure.

This is to learn how to organize code with model, view, and viewmodel.