Add buttons to change current cat using self
The binding for this was getting mixed up, so I assigned self to this to
avoid the mixup. It was causing issues with the change cat buttons, so I
chose to use self to fix it. Had to go through the other observables and
functions and change to self in the viewmodel.

Also had to move the with binding in the html div tag so it doesn't
include the buttons for the list of cats.